Full length links instead of just click on text
The navigation links were not properly clickable. We were supposed to click on the text only, to click on the link. The outer space around the link text is just unused. After this change, the links will be clickable in the whole li tag area instead of the a tag area.